Binary version of elf2link
The dataobjects are now good enough to create a binary version of elf2link:
script runtime on data.o: 0.35 s (real)
binary runtime on data.o: 0.011 s (debug, real)

data: A Data object that wraps around unsigned char *. This is better than the
      previous method of just holding a pointer to the raw memory as it
      insulates the rest of the code against a change in representation, and
      proides a convienent way to attach function to convert binary into hex
      representation

programHeader: Object representation of a elf program header (instructing the
               kernel as to what bits of the file to load into memory)

reloc: Object representation of an elf relocation entry. This is not yet fully
       supported in elf2link. (Support for libraries is to be implemented first)

General changes: Existing objects now provide a way of converting them into
                 LINK format

+++ libLink/elfHeader.cpp
@@ -7,7 +7,7 @@
 #endif
 
 using namespace std;
-ElfHeaderX86_64::ElfHeaderX86_64 (ElfReader &reader, int offset)
+ElfHeaderX86_64::ElfHeaderX86_64 (ElfReader &reader, long offset)
 {
     // over write the data in our data POD 
     reader.Read(0,&data,this->Size());
@@ -17,10 +17,10 @@ ElfHeaderX86_64::ElfHeaderX86_64 (string linkHeader)
     if ( false )  { // regex validation of linkHeader to prevent nastyness
         throw "Invalid LINK header: " + linkHeader;
     }
-    int nsections;
-    int nprog;
-    int nsyms;
-    int nrelocs;
+    long nsections;
+    long nprog;
+    long nsyms;
+    long nrelocs;
     readPtr = (char * ) & data;
     stringstream s(linkHeader);
     s >> nsections;
